requirements.txt
selenium
pymongo
Flask
python-dotenv
webdriver-manager
requests
gunicorn